* Remove KLIPS define in initiate.c
This define was supposed to limit some OE/on-demand code to KLIPS, but
there is really no reason anymore to do that as NETKEY can also do this.
As everyone on Linux enabled KLIPS support anyway, it was always there.

But now we also check/enable some initiate code when the currently used
stack is NETKEY.