* Added new option plutostderrlogtime= (default=no)
Add plutostderrlogtime= option to enable or disable logging the time
stamp before each log message (as syslog does). Some people want this
when logging to a file.  However, since the swan code is using this
output via plutostderrlog= in a file for the test suite, this needs to
be an option that can be turned off.