* XAUTH:  Use incoming XAUTH VID when picking best connection
I have prepared a patch witch solves for me following issue with Xauth
in Openswan.  Pluto may refuse to connect with a road warrior If some
misc connections (with and without Xauth) are configured. The reason is
that pluto do not regard Xauth policy in main_inI1_outR2 and may just
choose a not suitable connection for proceeding. In my patch I evaluate
XAUTH VID and use this information by connection finding.

Signed-off-by: Paul Wouters <paul@libreswan.org>